ref: Move user interactions from service to viewmodel

+++ POS/ViewModels/WarehouseFunctions/CreateDeliveryViewModel.cs
@@ -1,4 +1,5 @@
-﻿using System.Windows;
+﻿using System.Linq;
+using System.Windows;
 using POS.Utilities.RelayCommands;
 using POS.ViewModels.Base;
 using System.Windows.Input;
@@ -41,6 +42,15 @@ namespace POS.ViewModels.WarehouseFunctions
 
         private void DeleteIngredientFromDelivery(Ingredient ingredient)
         {
+            var ingredientToDelete = _deliveryService.DeliveryCollection.FirstOrDefault(i => i.Ingredient.IngredientId == ingredient.IngredientId);
+
+            if (ingredientToDelete!.Quantity == 1)
+            {
+                var result = MessageBox.Show("Czy usunąć składnik z listy całkowicie?", "", MessageBoxButton.YesNo, MessageBoxImage.Question);
+                if (result == MessageBoxResult.No)
+                    return;
+            }
+            
             _deliveryService.DeleteIngredientFromDeliveryCollection(ingredient);
         }
